added delete confirmation

+++ AppCalendar/views.py
@@ -63,6 +63,9 @@ def task(request, task_id=None):
 
 def task_delete(request, task_id=None):
     instance = Task.objects.get(pk=task_id)
-    instance.delete()
 
-    return HttpResponseRedirect(reverse('AppCalendar:calendar'))
+    if request.method == 'POST':
+        instance.delete()
+        return HttpResponseRedirect(reverse('AppCalendar:calendar'))
+
+    return render(request, 'AppCalendar/delete.html', {'task': instance})
